78 - Moves TableWithFilter to lib
