+++ libs/orchestrator-ui-components/src/lib/components/Table/TableWithFilter/index.ts
@@ -1,0 +1,1 @@
+export * from './TableWithFilter';
