+++ libs/orchestrator-ui-components/src/lib/components/Table/index.ts
@@ -9,3 +9,4 @@ export * from './constants';
 export * from './tableUtils';
 
 export * from './TableSettingsModal';
+export * from './TableWithFilter';
